Add a failing test for casting interface DateTimes
This commit adds a failing test showcasing what I believe to be a bug in
casting custom scalars (DateTime in this case) back to a Ruby DateTime
objects.

The current test strategy asserts that scalars defined on objects are
casted appropriately, but when a scalar is defined on an interface, it's
casted back as a string.