Fix an issue with loading fragments when file and directory names clash.
This fixes an issue where a folder like `app/views/issues/sidebar` would prevent us from loading the fragments defined in `app/views/issues/_sidebar.html.erb` (or `app/views/issues/sidebar.html.erb`).

As part of the changes in here, the meaning of the `path` attribute on `ViewModule`s was modified as well. It now no longer represents the file that a module was loaded from, but instead always is the "lookup path" for where to look for view files when trying to load a missing constant.