Immediately extend `ObjectClass` with class level readers
Metaclasses are lazily created, and extending the new class each time
will create the metaclass.  These classes seem to always bee instances
of `ObjectClass`, so if we add the class methods to `ObjectClass` then
we'll always have the correct class methods, but never need the
metaclass to be instantiated.

This saved about 250k allocations:

```
$ wc -l heap-before.json
 3280346 heap-before.json
$ wc -l heap-after.json
 3037087 heap-after.json
 ```