Add Release GitHub Actions workflow
This commit adds a GitHub Actions workflow that will build and publish a
new version of the gem to RubyGems. The workflow follows the pattern
described in RubyGems' "Trusted Publishing" documentation [1].

The workflow is triggered whenever a new Release is published by anyone
with permissions to do so on the project's GitHub repository. This
should work for pre-release versions, as well.

Releases may be created via the website [2] or by using the GitHub CLI:

```sh
gh release create v0.1.0 --generate-notes
```

Being sure to swap out `v0.1.0` with an appropriate (new or existing)
tag, of course. Note that draft releases and/or editing existing
releases will not re-trigger this workflow.

This commit follows up on my question/comment [3] on #3.

[1] https://guides.rubygems.org/trusted-publishing/releasing-gems/
[2] https://github.com/github-community-projects/graphql-client/releases
[3] https://github.com/github-community-projects/graphql-client/pull/3#issuecomment-1910521991