feat: Enable consumers to access the full response payload
Update the Response public API object to include full_response. this is non-breaking.

Passing the full response from the executing object to the client is currently done via the client checking with the object, versus it being part of the response,

In order to change this it requires changing the much more changes, as all the tests are using GraphQL::Schema for the execution of the graphql