fix: tests
Learned we can't get offenses directly off of the cop, it has to be
obtained from the commissioner.investigate method.

[rubocop sourcecode about not being able to access offenses off the cop base](https://github.com/rubocop/rubocop/blob/9c6e0ef9039f3602cc0b6e267b7be71ed1110d5b/lib/rubocop/cop/base.rb#L314-L318)

Learned we have to provide our cop as the forces argument to the
commissioner.investigate if we have our own investigate like we do for
the overfetch cop. If we don't it never runs. This is not the case with
the heredoc cop because it does not have an investigate method.

[rubocop commissioner source code where our custom cop invetigate method is actually called](https://github.com/rubocop/rubocop/blob/9c6e0ef9039f3602cc0b6e267b7be71ed1110d5b/lib/rubocop/cop/commissioner.rb#L85)
that calls [this function](https://github.com/rubocop/rubocop/blob/9c6e0ef9039f3602cc0b6e267b7be71ed1110d5b/lib/rubocop/cop/commissioner.rb#L163) and you'll notice the second argument is called cops

Signed-off-by: jmeridth <jmeridth@gmail.com>
Co-authored-by: Hannah Yiu <hkly@github.com>